Fixed bugs for SearchBar & Autocomplete View;  Added custom list for filters from View Controller

+++ GoogleMaps.framework/Versions/A/Headers/GMSOverlay.h
@@ -51,6 +51,6 @@
  * are effectively considered to be in a separate z-index group compared to
  * other overlays.
  */
-@property(nonatomic, assign) int zIndex;
+@property(nonatomic, assign) NSInteger zIndex;
 
 @end
